AI is now loads better Damage thresholds have been modified to be less receptive to very small blows, and more resistant to puncturing Dashing is now faster
NOTE: If you want to implement the new AI, add the prefab 'DangerWeapon' as a child object to the weapon gameobject. Also note that I haven't yet updated the Die() function, so you may receive an error after someone dies and their weapon still collides. This will be fixed soon.